+++ src/Screen.h
@@ -23,7 +23,7 @@ public:
 
 	// Screen implementation.
 	virtual void Init();
-	virtual bool Run(float delta) = 0;
+	virtual bool Run(double delta) = 0;
 	virtual void HandleInput(int key, int state, bool isMouseInput) = 0;
 
 	// Implement this if there's anything you want to get done outside of a destructor
